pyproject.toml
[tool.poetry]
name = "arsaga-stable-diffusion"
version = "0.1.0"
description = ""
authors = ["tatsuki.onoda <tatsuki.onoda@arsaga.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.12"
python-dotenv = "^1.0.1"
openai = "^1.14.0"
langchain-openai = "^0.0.8"
numpy = "^1.26.4"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pyright = "^1.1.357"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
